requirements.txt
# Google ADK and AI
google-adk>=0.5.0
google-genai>=1.0.0

# FastAPI Backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Async HTTP
httpx>=0.27.0
aiofiles>=23.2.0

# SSE Streaming
sse-starlette>=1.8.0

# File Processing
python-multipart>=0.0.6
reportlab>=4.0.0
Pillow>=10.0.0

# Data Processing
pandas>=2.0.0

# Utilities
python-dateutil>=2.8.0
uuid6>=2024.1.12
